+++ test/sources/data_schema/dataSchema.js
@@ -17,7 +17,7 @@ export default class DataSchema {
           field.type = this.getTypeOrArray(field.type)
         })
         if (!item.as || item.as === 'type') this.types[key] = Exonum.newType(item)
-        if (item.as === 'message') this.messages[key] = Exonum.newMessage(item)
+        if (item.as === 'message') this.messages[key] = Exonum.newTransaction(item)
       } else {
         if (item.as === 'array') {
           item.type = this.getTypeOrArray(item.type)